Refactor sampler constraints / filters into sampling.py
