+++ scripts/speech_recognition/estimate_token_bins.py
@@ -27,17 +27,15 @@ from omegaconf import OmegaConf
 
 from nemo.collections.asr.data.audio_to_text_lhotse import TokenizerWrapper
 from nemo.collections.common.data.lhotse.cutset import read_cutset_from_config
-from nemo.collections.common.data.lhotse.dataloader import (
+from nemo.collections.common.data.lhotse.dataloader import LhotseDataLoadingConfig, tokenize, tokenize_with_prompt
+from nemo.collections.common.data.lhotse.sampling import (
     DurationFilter,
     FixedBucketBatchSizeConstraint2D,
-    LhotseDataLoadingConfig,
     MultimodalFixedBucketBatchSizeConstraint2D,
     MultimodalSamplingConstraint,
     TokenCountFilter,
     TokenPerSecondFilter,
     TokenPerTokenFilter,
-    tokenize,
-    tokenize_with_prompt,
 )
 from nemo.collections.common.prompts.formatter import PromptFormatter
 from nemo.collections.common.tokenizers import AggregateTokenizer, SentencePieceTokenizer
